feat: required CID for accumulator witness data
This change validates that the accumulator data payload is a valid CID.
Additionally creates a Leaf type that is the return of the Get rpc call
instead of a vanilla tuple, (we still use a tuple over the wire).

I have a change pending against the SDK for these changes as well.

Fixes HES-383