refactor: changes credit representation to be multiple of attos
That is needed to avoid a decimal representation of `token_credit_rate`.
Suppose we want a `token_credit_rate` of 1 atto = 0.5 credit. For that we would need a `f64` type, or any other decimal representation, for `token_credit_rate`. That would lead to a bunch of problems.

By using attos for credits we can do 1 atoo = 500000000000000000 atto credits.

Signed-off-by: Bruno Calza <brunoangelicalza@gmail.com>