refactor: blob upload to use fixed tags instead of auto tags (#558)
Depends on https://github.com/recallnet/entanglement/tree/dig/tags

I don't think the tests fully cover this, but the shape of things should
be how I think this would be handled best given the existing apis.

- incoming blobs are tagged `temp-{hash}`
- the entangler tags all blobs they create as `ent-{hash}`
- after entanglement a hash sequence is generated and tagged
`temp-seq-{hash}`
- the `temp-` and `ent-` tags relevant to this sequence are deleted
- when the chain upgrades the data, it removes `temp-seq-{hash}` and
adds `stored-{hash}`


## Open Questions

- The system currently will fail/race in the situation two blobs are
uploaded at the same time, as the tagging is not scoped to a specific
upload. This could be solved by introducting another identifier in the
tag, or tracking uploads in progress more explicitly somehow
- The way the chain interacts with this, assumes that there is a single
tag per stored object, what about the same object being referenced
multiple times on chain?
- TTLs for uploaded objects are not yet implemented, this will need to
be done in it's own PR
- Currently this uses a simple string based format for tagging, but tags
can be arbitrary bytes, so an optimization will be to introduce an
actual structural type that gets encoded into the tag name, making this
more robust, and easier to extend in the future