fix: preserve filetype/extension when downloading an object (#576)
# Summary

Closes https://github.com/recallnet/js-recall/issues/101

This PR adds the header `Content-Disposition` to the Objects API when
downloading the object. It checks if the object's key has an extension,
if so, it uses the object's key as the Content-Disposition's `filename`.
If not, tries to guess based on content type.

# Notes
- This PR might fix Brave browser issue, but it will probably change the
current behavior of other browsers. That is because the default
`Content-Disposition` is `inline` and not `attachment`. So, before going
forward with this we must decide on what behavior we want.
- It is a bit hard to test now because of the download problem 
- I noticed we could improve the CLI to also check the extension to
guess the content type. See
https://github.com/recallnet/rust-recall/pull/235. Lmk if that makes
sense

Signed-off-by: Bruno Calza <brunoangelicalza@gmail.com>