chore: upgrade to iroh v0.35, update entangler dep (#626)
depends on [entangler
#45](https://github.com/recallnet/entanglement/pull/45), after merging
I'll update version deps to match the new version on engangler/main.

targets #625, this updates iroh to version 0.35, which will match the
version of the deployed relays.

This PR also incorporates the streaming updates & config constant change
that drops the `ENTANGLER_P` param, which should be reviewed.